Add files via upload
We will design for the application of IOT in manufacturing industry where pressure, temperature and humidity must be monitored every now and then. So we use three sensors that outputs values at different times and visualizations are analysed in design centre. So the sensors in shop floor of the factory can be kept under check by observing the values at the design centre. The hardware in this application is on sensors and values are transformed via MQTT server. The sensor position, values at different times for three sensors are observed without requiring going to shop floor. So person in one department can observe the values of sensors situated in other place. This is the essence of using IOT here. We have used mqtt to publish the sensor values and then subscribe the values at the receiving end. We then visualize using different plots to understand danger levels.